Turn generator function into class in scraper.py
The reason for doing this is simply to be able to do len(accommodations)
in main.py instead of doing next(accommodations) to get the length.
Yielding an int as the first value (the length) instead of
Accommodation as it does otherwise always seemed a bit weird to me.

The new code has few extra lines and is a bit harder to read, but I
believe it's worth the cleaner syntax when using it.