Move 0 arguments check just before parsing
This is basically a revert of some other commit I did once. The reason
I changed it before was that checking the arguments should be done
first because no parsing would be necessary.

My reasoning for moving it back down again is that it is just more
logical to be doing all parsing in one place, instead of doing this
check at the top and the actual parsing at the bottom. The loss in
performance in negligible because constructing the parser is rather
fast.