Allow filters in list command
This is quite and easy change to motivate, filtering what
accommodations to see is quite useful. In order to achieve this kwargs
argument from a few functions had to be changed into regular arguments,
which is a good thing. So I believe that this change even made the code
better as a whole.